--- src/worker/Params.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-
-class Params(BaseModel):
-    filename: str
-    namespace: str
