--- src/worker/Task.py
@@ -1,7 +1,0 @@
-from pydantic import BaseModel
-from Params import Params
-
-
-class Task(BaseModel):
-    task: str
-    params: Params
