+++ src/worker/queue_processor.py
@@ -7,11 +7,9 @@ from rsmq.consumer import RedisSMQConsumer
 from rsmq import RedisSMQ
 from sentry_sdk.integrations.redis import RedisIntegration
 import sentry_sdk
-from . import Message
 
-from Task import Task
-from convert_to_pdf import convert_to_pdf
-
+from .models import Message, Task
+from .convert_to_pdf import convert_to_pdf
 
 SERVICE_NAME = "convert-to-pdf"
 RESULTS_QUEUE_NAME = f"{SERVICE_NAME}_results"
